+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Extend type Options to allow for x/y use
+
 ## [2.2.3] - 2022-12-12
 
 - Add support for Cypress 12.x.